fixed example
