+++ examples/SubMenu/SubMenu.ino
@@ -54,7 +54,7 @@ MenuItem mainMenu[] = {ItemHeader(),
 /**
  * Create submenu and precise its parent
  */
-MenuItem settingsMenu[] = {ItemSubHeader(mainMenu),
+MenuItem settingsMenu[] = {ItemHeader(mainMenu),
                            MenuItem("Backlight"),
                            MenuItem("Contrast"),
                            ItemFooter()};
